requirements.txt
fastapi==0.70.0
uvicorn==0.15.0
pymongo[srv]==4.1.1
pydantic==1.9.0
uuid==1.30
starlette==0.16.0
passlib[bcrypt]==1.7.4
pyjwt==2.3.0
rabbitpy==2.0.1
requests==2.26.0
datetime==4.4
httpx==0.21.1
